Refactor to com.actualplatform.android.activation.development
