+++ blinkreceipt-demo/ActualActivations/.gitignore
