+++ blinkreceipt-demo/ActualActivations/proguard-rules.pro
