+++ blinkreceipt-demo/ActualActivations/src/main/AndroidManifest.xml
