+++ blinkreceipt-demo/ActualActivations/src/main/java/com/actualplatform/android/activation/development/BlinkReceiptApplication.kt
@@ -1,7 +1,7 @@
-package com.blinkreceipt.engage
+package com.actualplatform.android.activation.development
 
 import android.app.Application
-import com.blinkreceipt.engage.ui.ActivationsActivity
+import com.actualplatform.android.activation.development.ui.ActivationsActivity
 import com.microblink.BlinkReceiptSdk
 import com.microblink.core.InitializeCallback
 import com.microblink.logcat.LogcatManager
@@ -13,7 +13,7 @@ class BlinkReceiptApplication: Application() {
 
         BlinkReceiptSdk.initialize(this, object : InitializeCallback {
             override fun onComplete() {
-                ActivationsActivity.applySettings(this@BlinkReceiptApplication)
+                ActivationsActivity.Companion.applySettings(this@BlinkReceiptApplication)
             }
 
             override fun onException(throwable: Throwable) {
