+++ blinkreceipt-demo/ActualActivations/src/main/java/com/actualplatform/android/activation/development/MainActivity.kt
@@ -1,6 +1,7 @@
-package com.blinkreceipt.engage
+package com.actualplatform.android.activation.development
 
 import android.content.Intent
+import android.hardware.camera2.CameraCharacteristics
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.LocalActivity
@@ -14,7 +15,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import com.blinkreceipt.engage.ui.ActivationsActivity
+import com.actualplatform.android.activation.development.ui.ActivationsActivity
 import com.microblink.ScanOptions
 import com.microblink.camera.ui.CameraCharacteristics
 
@@ -40,13 +41,13 @@ private fun EngageHomeScreen() {
                 activity?.let {
                     val intent = Intent(context, ActivationsActivity::class.java).apply {
                         putExtra(
-                            ActivationsActivity.OPTIONS_KEY,
+                            ActivationsActivity.Companion.OPTIONS_KEY,
                             ScanOptions.Builder()
                                 .detectDuplicates(true)
                                 .build(),
                         )
                         putExtra(
-                            ActivationsActivity.CAMERA_CHARACTERISTICS,
+                            ActivationsActivity.Companion.CAMERA_CHARACTERISTICS,
                             CameraCharacteristics.Builder()
                                 .cameraPermission(true)
                                 .build(),
