+++ blinkreceipt-demo/ActualActivations/src/main/java/com/actualplatform/android/activation/development/ui/DebugPlacementsSection.kt
@@ -1,4 +1,4 @@
-package com.blinkreceipt.engage.ui
+package com.actualplatform.android.activation.development.ui
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -12,8 +12,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.actualplatform.engage.models.PlacementLayout
-import com.blinkreceipt.engage.R
+import com.actualplatform.android.models.PlacementLayout
+import com.actualplatform.android.R
 
 @Composable
 internal fun DebugPlacementsSection(
