+++ blinkreceipt-demo/ActualActivations/src/main/java/com/actualplatform/android/activation/development/ui/SettingsComponents.kt
@@ -1,4 +1,4 @@
-package com.blinkreceipt.engage.ui
+package com.actualplatform.android.activation.development.ui
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
