+++ blinkreceipt-demo/ActualActivations/src/main/res/drawable/ic_arrow_back.xml
