+++ blinkreceipt-demo/ActualActivations/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Blink Receipt Engage</string>
+    <string name="app_name">Actual Activations</string>
 
     <string name="activations_title">Activations</string>
     <string name="engage_settings_title">Settings</string>
