+++ blinkreceipt-demo/settings.gradle
@@ -5,7 +5,7 @@ include ':BlinkReceiptEarnings'
 include ':BlinkReceiptImap'
 include ':BlinkReceiptOutlook'
 include ':BlinkReceiptStaticScan'
-include ':BlinkReceiptEngage'
+include ':ActualActivations'
 include ':BlinkReceiptCustomScan'
 include ':BlinkReceiptSurveys'
 include ':BlinkReceiptAccountLinking'
